- Merged GitBlamer and GitForensics test into one - fixed javadoc warnings
